requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
azure-identity
azure-core
msal
